[PORT]Support deep data binding for some properties (#2552)
* init

* init

* add tests

* fix comments

* format code

* fix comments

* retrigger

* retrigger
