+++ libraries/botbuilder-dialogs-adaptive-testing/tests/resources/ActionTests/Action_SetProperties.test.dialog
@@ -26,6 +26,18 @@
                             {
                                 "property": "$p4",
                                 "value": "='billybob'"
+                            },
+                            {
+                                "property": "user.profile.userName",
+                                "value": "jack"
+                            },
+                            {
+                                "property": "user.profile",
+                                "value":
+                                    {
+                                        "name": "${sentenceCase(user.profile.userName)}",
+                                        "age": "${10 * 2}"
+                                    }
                             }
                         ]
                     },
@@ -42,6 +54,10 @@
                     },                    {
                         "$kind": "Microsoft.SendActivity",
                         "activity": "${$p4}"
+                    },
+                    {
+                        "$kind": "Microsoft.SendActivity",
+                        "activity": "${user.profile.name}"
                     }
                 ]
             }
@@ -66,6 +82,10 @@
         {
             "$kind": "Microsoft.Test.AssertReply",
             "text": "billybob"
+        },
+        {
+            "$kind": "Microsoft.Test.AssertReply",
+            "text": "Jack"
         }
     ]
 }
