+++ libraries/botbuilder-dialogs-adaptive-testing/tests/resources/ActionTests/Action_SetProperty.test.dialog
@@ -70,6 +70,23 @@
           {
             "$kind": "Microsoft.SendActivity",
             "activity": "${user.$instance.name}"
+          },
+          {
+            "$kind": "Microsoft.SetProperty",
+            "property": "user.profile.userName",
+            "value": "jack"
+          },
+          {
+            "$kind": "Microsoft.SetProperty",
+            "property": "user.profile",
+            "value": {
+                "name": "${sentenceCase(user.profile.userName)}",
+                "age": "${10 * 2}"
+            }
+          },
+          {
+            "$kind": "Microsoft.SendActivity",
+            "activity": "${user.profile.name}"
           }
         ]
       }
@@ -90,6 +107,10 @@
     {
       "$kind": "Microsoft.Test.AssertReply",
       "text": "name"
+    },
+    {
+      "$kind": "Microsoft.Test.AssertReply",
+      "text": "Jack"
     }
   ]
 }
