+++ libraries/botbuilder-dialogs-adaptive/src/actions/endDialog.ts
@@ -7,6 +7,7 @@
  */
 import { DialogTurnResult, DialogContext, Dialog } from 'botbuilder-dialogs';
 import { ValueExpression, BoolExpression } from 'adaptive-expressions';
+import { replaceJsonRecursively } from '../jsonExtensions';
 
 export class EndDialog<O extends object = {}> extends Dialog<O> {
     /**
@@ -35,7 +36,12 @@ export class EndDialog<O extends object = {}> extends Dialog<O> {
         }
 
         if (this.value) {
-            const value = this.value.getValue(dc.state);
+            let value = this.value.getValue(dc.state);
+            
+            if (value) {
+                value = replaceJsonRecursively(dc.state, value);
+            }
+
             return await this.endParentDialog(dc, value);
         }
