+++ libraries/botbuilder-dialogs-adaptive/src/actions/setProperties.ts
@@ -9,6 +9,7 @@ import { StringUtils } from 'botbuilder-core';
 import { Dialog, DialogContext, DialogTurnResult } from 'botbuilder-dialogs';
 import { Converter } from 'botbuilder-dialogs-declarative';
 import { ValueExpression, StringExpression, BoolExpression } from 'adaptive-expressions';
+import { replaceJsonRecursively } from '../jsonExtensions';
 
 export interface PropertyAssignment {
     property: StringExpression;
@@ -49,7 +50,12 @@ export class SetProperties<O extends object = {}> extends Dialog<O> {
 
         for (let i = 0; i < this.assignments.length; i++) {
             const assignment = this.assignments[i];
-            const value = assignment.value.getValue(dc.state);
+            let value = assignment.value.getValue(dc.state);
+            
+            if (value) {
+                value = replaceJsonRecursively(dc.state, value);
+            }
+
             const property = assignment.property.getValue(dc.state);
             dc.state.setValue(property, value);
         }
